// backend/src/services/secret-sync/secret-sync-types.ts
export type TSyncOptionsConfig = {
  canImportSecrets: boolean;
  canRemoveSecretsOnDeletion?: boolean;
  supportsKeySchema?: boolean;
  supportsDisableSecretDeletion?: boolean;
};